+++ banking-bench/Cargo.toml
@@ -21,6 +21,7 @@ solana-ledger = { path = "../ledger", version = "=1.8.0" }
 solana-logger = { path = "../logger", version = "=1.8.0" }
 solana-measure = { path = "../measure", version = "=1.8.0" }
 solana-perf = { path = "../perf", version = "=1.8.0" }
+solana-poh = { path = "../poh", version = "=1.8.0" }
 solana-runtime = { path = "../runtime", version = "=1.8.0" }
 solana-streamer = { path = "../streamer", version = "=1.8.0" }
 solana-sdk = { path = "../sdk", version = "=1.8.0" }
