+++ core/Cargo.toml
@@ -22,7 +22,6 @@ bv = { version = "0.11.1", features = ["serde"] }
 bs58 = "0.4.0"
 byteorder = "1.3.4"
 chrono = { version = "0.4.11", features = ["serde"] }
-core_affinity = "0.5.10"
 crossbeam-channel = "0.5"
 ed25519-dalek = "=1.0.1"
 fs_extra = "1.2.0"
@@ -60,10 +59,11 @@ solana-gossip = { path = "../gossip", version = "=1.8.0" }
 solana-ledger = { path = "../ledger", version = "=1.8.0" }
 solana-logger = { path = "../logger", version = "=1.8.0" }
 solana-merkle-tree = { path = "../merkle-tree", version = "=1.8.0" }
-solana-metrics = { path = "../metrics", version = "=1.8.0" }
 solana-measure = { path = "../measure", version = "=1.8.0" }
+solana-metrics = { path = "../metrics", version = "=1.8.0" }
 solana-net-utils = { path = "../net-utils", version = "=1.8.0" }
 solana-perf = { path = "../perf", version = "=1.8.0" }
+solana-poh = { path = "../poh", version = "=1.8.0" }
 solana-program-test = { path = "../program-test", version = "=1.8.0" }
 solana-rpc = { path = "../rpc", version = "=1.8.0" }
 solana-runtime = { path = "../runtime", version = "=1.8.0" }
@@ -73,7 +73,6 @@ solana-frozen-abi-macro = { path = "../frozen-abi/macro", version = "=1.8.0" }
 solana-stake-program = { path = "../programs/stake", version = "=1.8.0" }
 solana-storage-bigtable = { path = "../storage-bigtable", version = "=1.8.0" }
 solana-streamer = { path = "../streamer", version = "=1.8.0" }
-solana-sys-tuner = { path = "../sys-tuner", version = "=1.8.0" }
 solana-transaction-status = { path = "../transaction-status", version = "=1.8.0" }
 solana-version = { path = "../version", version = "=1.8.0" }
 solana-vote-program = { path = "../programs/vote", version = "=1.8.0" }
@@ -111,9 +110,6 @@ name = "gen_keys"
 
 [[bench]]
 name = "sigverify_stage"
-
-[[bench]]
-name = "poh"
 
 [[bench]]
 name = "retransmit_stage"
