+++ poh/Cargo.toml
@@ -1,0 +1,39 @@
+[package]
+name = "solana-poh"
+version = "1.8.0"
+description = "Solana PoH"
+authors = ["Solana Maintainers <maintainers@solana.foundation>"]
+repository = "https://github.com/solana-labs/solana"
+license = "Apache-2.0"
+homepage = "https://solana.com/"
+documentation = "https://docs.rs/solana-poh"
+edition = "2018"
+
+[dependencies]
+core_affinity = "0.5.10"
+crossbeam-channel = "0.5"
+log = "0.4.11"
+solana-ledger = { path = "../ledger", version = "=1.8.0" }
+solana-measure = { path = "../measure", version = "=1.8.0" }
+solana-metrics = { path = "../metrics", version = "=1.8.0" }
+solana-runtime = { path = "../runtime", version = "=1.8.0" }
+solana-sdk = { path = "../sdk", version = "=1.8.0" }
+solana-sys-tuner = { path = "../sys-tuner", version = "=1.8.0" }
+thiserror = "1.0"
+
+[dev-dependencies]
+bincode = "1.3.1"
+matches = "0.1.6"
+rand = "0.7.0"
+solana-logger = { path = "../logger", version = "=1.8.0" }
+solana-perf = { path = "../perf", version = "=1.8.0" }
+
+[lib]
+crate-type = ["lib"]
+name = "solana_poh"
+
+[[bench]]
+name = "poh"
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
