+++ poh/benches/poh.rs
@@ -3,8 +3,8 @@
 #![feature(test)]
 extern crate test;
 
-use solana_core::poh_service::DEFAULT_HASHES_PER_BATCH;
 use solana_ledger::poh::Poh;
+use solana_poh::poh_service::DEFAULT_HASHES_PER_BATCH;
 use solana_sdk::hash::Hash;
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::{Arc, Mutex};
