+++ poh/src/lib.rs
@@ -1,0 +1,10 @@
+#![allow(clippy::integer_arithmetic)]
+pub mod poh_recorder;
+pub mod poh_service;
+
+#[macro_use]
+extern crate solana_metrics;
+
+#[cfg(test)]
+#[macro_use]
+extern crate matches;
